BJ-902: Refactor to introduce AbstractCacheService
