+++ src/main/java/com/rbkmoney/threeds/server/handle/constraint/pres/cardrangedata/CardRangeContentConstraintValidationHandlerImpl.java
@@ -5,7 +5,7 @@ import com.rbkmoney.threeds.server.domain.root.emvco.PRes;
 import com.rbkmoney.threeds.server.dto.ConstraintType;
 import com.rbkmoney.threeds.server.dto.ConstraintValidationResult;
 import com.rbkmoney.threeds.server.handle.constraint.pres.PResConstraintValidationHandler;
-import com.rbkmoney.threeds.server.service.CacheService;
+import com.rbkmoney.threeds.server.service.cache.CacheService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
