+++ src/main/java/com/rbkmoney/threeds/server/handle/constraint/rreq/RReqRequiredContentConstraintValidationHandlerImpl.java
@@ -7,7 +7,7 @@ import com.rbkmoney.threeds.server.domain.message.MessageCategory;
 import com.rbkmoney.threeds.server.domain.root.emvco.RReq;
 import com.rbkmoney.threeds.server.dto.ConstraintValidationResult;
 import com.rbkmoney.threeds.server.dto.RReqTransactionInfo;
-import com.rbkmoney.threeds.server.service.CacheService;
+import com.rbkmoney.threeds.server.service.cache.CacheService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
